publish algo
Abstract—In-network caching and flexible routing are two of the
most celebrated advantages of next generation network infrastructures. Yet few solutions are available for jointly optimizing caching
and routing that provide performance guarantees for networks
with arbitrary topology. We take a holistic approach towards this
fundamental problem by analyzing its complexity in all the cases
and developing polynomial-time algorithms with approximation
guarantees in important special cases. We also reveal the fundamental challenge in achieving guaranteed approximation in the
general case and propose an alternating optimization algorithm
with good empirical performance and fast convergence. Our algorithms have demonstrated superior performance in both routing
cost and congestion compared to the state-of-the-art solutions in
evaluations based on real topology and request traces.